requirements.txt
streamlit
langchain
openai
cassio
PyPDF2
tiktoken
python-dotenv